rename script files
